Uploaded Framework and Initial Menu
Here is the initial framework, with graphics, and audio (that should be
changed soon, it isn't very fitting).  Please download the current file,
and run it.  As long there are no issues, the graphics load (just a 2D
splash screen at the moment), and the audio plays, then everything is in
order and we will be able to move forward.  If you have any issues
running this code, then let me know and I'll get it fixed.  From here I
am going to update a few milestones, and start setting up github for
this project.