Added Intro game state
Added a game state for the intro.  When you select "New Game" from the
title menu, it displays a splash screen with Castle Baron.  Text slowly
scrolls up from the bottom of the screen in order to tell the game's
story.  There is also a button in the top right corner that can be used
to skip the intro scene and go to the overworld.